+++ lib/project_types/script/layers/infrastructure/languages/project_creator.rb
@@ -26,7 +26,6 @@ module Script
           )
 
             project_creators = {
-              "assemblyscript" => AssemblyScriptProjectCreator,
               "typescript" => TypeScriptProjectCreator,
               "wasm" => WasmProjectCreator,
             }