+++ test/project_types/script/layers/domain/extension_point_test.rb
@@ -7,9 +7,6 @@ describe Script::Layers::Domain::ExtensionPoint do
   let(:config) do
     {
       "libraries" => {
-        "assemblyscript" => {
-          "package" => "@shopify/extension-point-as-fake",
-        },
         "typescript" => {
           "package" => "@shopify/extension-point-ts-fake",
         },
@@ -19,9 +16,6 @@ describe Script::Layers::Domain::ExtensionPoint do
   let(:config_with_rust) do
     {
       "libraries" => {
-        "assemblyscript" => {
-          "package" => "@shopify/extension-point-as-fake",
-        },
         "typescript" => {
           "package" => "@shopify/extension-point-ts-fake",
         },
@@ -102,8 +96,8 @@ describe Script::Layers::Domain::ExtensionPoint do
         subject { Script::Layers::Domain::ExtensionPoint.new(type, config_with_rust) }
         it "should return all the implemented libraries" do
           extension_point = subject
-          assert_equal 3, extension_point.libraries.all.count
-          refute_nil extension_point.libraries.for("assemblyscript")
+          assert_equal 2, extension_point.libraries.all.count
+          refute_nil extension_point.libraries.for("typescript")
           refute_nil extension_point.libraries.for("rust")
         end
       end
@@ -114,7 +108,7 @@ describe Script::Layers::Domain::ExtensionPoint do
         it "should not return that library" do
           extension_point = subject
 
-          assert_equal 2, extension_point.libraries.all.count
+          assert_equal 1, extension_point.libraries.all.count
           assert_nil extension_point.libraries.for("rust")
         end
       end
@@ -128,8 +122,7 @@ describe Script::Layers::Domain::ExtensionPoint do
         let(:include_betas) { true }
 
         it "returns all the languages of the libraries" do
-          assert_equal 3, subject.count
-          assert_includes subject, "assemblyscript"
+          assert_equal 2, subject.count
           assert_includes subject, "typescript"
           assert_includes subject, "rust"
         end
@@ -139,8 +132,7 @@ describe Script::Layers::Domain::ExtensionPoint do
         let(:include_betas) { false }
 
         it "returns only non-beta languages of the libraries" do
-          assert_equal 2, subject.count
-          assert_includes subject, "assemblyscript"
+          assert_equal 1, subject.count
           assert_includes subject, "typescript"
         end
       end