+++ CHANGELOG.md
@@ -1,9 +1,12 @@
 From version 2.6.0, the sections in this file adhere to the [keep a changelog](https://keepachangelog.com/en/1.0.0/) specification.
 
 ## [Unreleased]
-### Fixed
-* [#2088](https://github.com/Shopify/shopify-cli/pull/2088): Update theme-check to 1.10.1
+### Added
+* [#2103](https://github.com/Shopify/shopify-cli/pull/2103): Improve `shopify theme package` to include the `release-notes.md` file
 
+### Fixed
+* [#2112](https://github.com/Shopify/shopify-cli/pull/2112): Fix intermittent error ("can't add a new key into hash during iteration") in the `theme push` command
+* [#2088](https://github.com/Shopify/shopify-cli/pull/2088): Update theme-check to 1.10.1
 
 ## Version 2.13.0
 
@@ -14,6 +17,9 @@ From version 2.6.0, the sections in this file adhere to the [keep a changelog](h
 ### Fixed
 * [#2092](https://github.com/Shopify/shopify-cli/pull/2092): Fix `RootHelper` parse logic to support options with an equal (e.g.: `option=value`)
 * [#2089](https://github.com/Shopify/shopify-cli/pull/2089): Use javy version 0.2.0
+
+### Removed
+* [#2102](https://github.com/Shopify/shopify-cli/pull/2102): Remove AssemblyScript as a supported script language.
 
 ## Version 2.12.0
 ### Added